bug with fast running fixed

+++ client/src/actor/actor.cpp
@@ -11,6 +11,6 @@ anti_stun::order actor::process_anti_stun() {
   return anti_stun.process(settings.anti_stun);
 }
 
-void actor::process_fast_run(double& out_run_speed) {
+void actor::process_fast_run(float& out_run_speed) {
   fast_run.process(settings.fast_run, out_run_speed);
 }
